Remove Bob the String-Builder
Seeing as dynarr.h already exists, there is no real use for Bob the
String-Builder.  It was *a bit* nice because of the generic u8strpush()
macro, but it should be very easy to implement yourself, and you only
really ever need 1 or 2 specific functions that are a few lines of code
each.